cloud_functions/upload_to_hf/requirements.txt
functions-framework==3.*
google-cloud-storage
datasets
huggingface_hub
torchvision